requirements.txt
# Core CV & DL
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Object Detection
ultralytics>=8.0.0

# OCR
easyocr>=1.7.0

# Data Augmentation
albumentations>=1.3.0

# API
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Demo UI
streamlit>=1.28.0

# Utilities
scikit-learn>=1.3.0
matplotlib>=3.7.0
tqdm>=4.65.0
pydantic>=2.0.0
